requirements.txt
asgiref==3.5.2
astroid==2.12.12
autopep8==2.0.0
boto3==1.26.3
botocore==1.29.3
certifi==2022.9.24
charset-normalizer==2.1.1
coreapi==2.3.3
coreschema==0.0.4
defusedxml==0.7.1
diff-match-patch==20200713
dill==0.3.6
disposable-email-domains==0.0.84
Django==4.1.3
django-ckeditor==6.5.1
django-cors-headers==3.13.0
django-filter==22.1
django-import-export==3.0.2
django-js-asset==2.0.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-yasg==1.21.4
et-xmlfile==1.1.0
idna==3.4
inflection==0.5.1
isort==5.10.1
itypes==1.2.0
Jinja2==3.1.2
jmespath==1.0.1
kavenegar==1.1.2
lazy-object-proxy==1.8.0
MarkupPy==1.14
MarkupSafe==2.1.1
mccabe==0.7.0
odfpy==1.4.1
openpyxl==3.0.10
packaging==21.3
Pillow==9.3.0
platformdirs==2.5.3
psycopg2-binary==2.9.5
pycodestyle==2.9.1
PyJWT==2.6.0
pylint==2.15.5
pyotp==2.7.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2022.6
PyYAML==6.0
requests==2.28.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
s3transfer==0.6.0
six==1.16.0
sqlparse==0.4.3
tablib==3.3.0
tomli==2.0.1
tomlkit==0.11.6
typing_extensions==4.4.0
uritemplate==4.1.1
urllib3==1.26.12
wrapt==1.14.1
xlrd==2.0.1
xlwt==1.3.0
